Corrected genre and list all genre
